[APB-3635][MP] Adding matching get routes to posts to escape error pages trying to GET a POST endpoint (#514)
This adds matching GET routes for non-trivial POST endpoints. If a POST endpoint fails, the POST URL is left shown in the browser address bar. Clicking the browser refresh would re-attempt the POST, but pressing enter in the address bar would make a GET request to the POST endpoint's URL.
This change supports users who try to refresh an error page by pressing enter in the browser address bar (#514)